Split mcp23018 and keyboard encoder support
