+++ keyboards/torn/torn.h
@@ -36,6 +36,3 @@
 // clang-format on
 
 void torn_set_led(uint8_t led, bool on);
-
-void secondary_encoder_read(void);
-void secondary_encoder_init(void);
